% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments-calculate.R
\name{calculateFragments}
\alias{calculateFragments}
\alias{modificationPositions}
\alias{defaultNeutralLoss}
\alias{calculateFragments,character,missing-method}
\title{Calculate ions produced by fragmentation with variable modifications}
\usage{
\S4method{calculateFragments}{character,missing}(
  sequence,
  type = c("b", "y"),
  z = 1,
  fixed_modifications = NULL,
  variable_modifications = NULL,
  max_mods = Inf,
  neutralLoss = defaultNeutralLoss(),
  verbose = TRUE
)
}
\arguments{
\item{sequence}{\code{character()} providing a peptide sequence. If positional
modifications are included in the sequence, variable modifications may not be
used. See examples below for more detail.}

\item{type}{\code{character} vector of target ions; possible values:
\code{c("a", "b", "c", "x", "y", "z")}. Default is \code{type = c("b", "y")}.}

\item{z}{numeric with a desired charge state; default is 1.}

\item{fixed_modifications}{Deprecated parameter. Please use
\code{\link[PTMods:addFixedModifications]{PTMods::addFixedModifications()}} to generate sequences with positional
modifications instead. Named \code{numeric} or \code{character}. If a \code{character} is
given, values must be in UniMod name or UniMod ID format
(e.g. \code{"Phospho"}, \code{"UNIMOD:21"}). The annotation style of
the values is preserved in the output. Specifies which fixed
modifications are applied to which amino acids.}

\item{variable_modifications}{Deprecated parameter. Please use
\code{\link[PTMods:addVariableModifications]{PTMods::addVariableModifications()}} to generate sequences with positional
modifications instead. Named \code{numeric} or \code{character}. If a \code{character}
is given, values must be in UniMod name or UniMod ID format
(e.g. \code{"Phospho"}, \code{"UNIMOD:21"}). The annotation style of
the values is preserved in the output. Specifies which
variable modifications are used on which amino acids.}

\item{max_mods}{Deprecated parameter. Please use in combination with
PTMods::addVariableModifications() instead. A numeric indicating the
maximum number of variable modifications
allowed on the sequence at once. Does not include fixed modifications.
Default value is positive infinity.}

\item{neutralLoss}{\code{list}, it has to have two named elments,
namely \code{water} and \code{ammonia} that contain a \code{character} vector
which type of neutral loss should be calculated.  Currently
neutral loss on the C terminal \code{"Cterm"}, at the amino acids
\code{c("D", "E", "S", "T")} for \code{"water"} (shown with an \verb{_}) and
\code{c("K", "N", "Q", "R")} for \code{"ammonia"} (shown with an \code{*}) are
supported.

\if{html}{\out{<div class="sourceCode">}}\preformatted{There is a helper function `defaultNeutralLoss()` that returns
the correct list. It has two arguments `disableWaterLoss` and
`disableAmmoniaLoss` to remove single neutral loss options. See
the example section for use cases.
}\if{html}{\out{</div>}}}

\item{verbose}{\code{logical(1)}. Deprecated parameter.
If \code{TRUE} (default) the used modifications are printed.}
}
\value{
A \code{data.frame} showing all the
ions produced by fragmentation with all possible combinations of modifications.
The used variable modifications are displayed in the \code{peptide} column through the
use of amino acids followed by the modification within brackets.
Fixed modifications are not displayed.
}
\description{
This method calculates a-, b-, c-, x-, y- and z-ions produced by
fragmentation.

Available methods
\itemize{
\item The default method with signature \code{sequence = "character"} and
\code{object = "missing"} calculates the theoretical fragments for a
peptide sequence. It returns a \code{data.frame} with the columns
\code{mz}, \code{ion}, \code{type}, \code{pos}, \code{z}, \code{seq} and \code{peptide}.
\item Additional method can be defined that will adapt their behaviour
based on spectra defined in \code{object}. See for example the MSnbase
package that implements a method for objects of class
\code{Spectrum2}.
}
}
\examples{

## No modifications
calculateFragments("ACE")

## Multiple ion types and charge states
calculateFragments("ACE",
                   type = c("a", "b", "c", "x", "y", "z"),
                   z = 1:2)

## Positional modification written directly in the sequence string
## The annotation style must be supported by PTMods::convertAnnotation
calculateFragments("T[+79.966]CE")
calculateFragments("T[Phospho]CE")
calculateFragments("T[UNIMOD:21]C[Carbamidomethyl]E")

## neutral loss
defaultNeutralLoss()

## disable water loss on the C terminal
defaultNeutralLoss(disableWaterLoss="Cterm")

## real example
calculateFragments("PQR")
calculateFragments("PQR",
                   neutralLoss=defaultNeutralLoss(disableWaterLoss="Cterm"))
calculateFragments("PQR",
                   neutralLoss=defaultNeutralLoss(disableAmmoniaLoss="Q"))

## disable neutral loss completely
calculateFragments("PQR", neutralLoss=NULL)

## Recommended workflow: use PTMods functions to produce positional sequences
## before calling calculateFragments.

## Fixed modification (Carbamidomethyl on C) using addFixedModifications
seq_fixed <- PTMods::addFixedModifications("ACE",
                                           fixedModifications = c(C = 57.02))
calculateFragments(seq_fixed)

## Fixed modification including N-terminus using addFixedModifications
seq_nterm <- PTMods::addFixedModifications(
    "ACE",
    fixedModifications = c(C = 57.02, Nterm = 229.16))
calculateFragments(seq_nterm)

## Variable modification (delta mass on A) using addVariableModifications
seq_var <- PTMods::addVariableModifications("ACE",
                                            variableModifications = c(A = 43.25))
calculateFragments(seq_var)

## Both fixed and variable modifications using addModifications
seq_mods <- PTMods::addModifications("ARGSHKATC",
                                     fixedModifications = c(C = 57),
                                     variableModifications = c(S = 79, T = 79),
                                     maxMods = 2)
calculateFragments(seq_mods)
}
\author{
Sebastian Gibb \href{mailto:mail@sebastiangibb.de}{mail@sebastiangibb.de}

Guillaume Deflandre \href{mailto:guillaume.deflandre@uclouvain.be}{guillaume.deflandre@uclouvain.be}
}
