% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments-calculate2.R
\name{calculateFragments2}
\alias{calculateFragments2}
\alias{calculateFragments2,}
\alias{modificationPositions,}
\alias{cumsumFragmentMasses,}
\alias{character,missing-method}
\title{Calculate ions produced by fragmentation, including possible modifications}
\usage{
calculateFragments2(
  sequence,
  type = c("b", "y"),
  z = 1,
  fixed_modifications = c(C = 57.02146),
  variable_modifications = NULL,
  max_mods = Inf,
  neutralLoss = defaultNeutralLoss(),
  verbose = TRUE
)
}
\arguments{
\item{sequence}{character() providing a peptide sequence.}

\item{type}{character vector of target ions; possible values:
c("a", "b", "c", "x", "y", "z"). Default is type = c("b", "y").}

\item{z}{numeric with a desired charge state; default is 1.}

\item{fixed_modifications}{A named numeric vector of used fixed modifications.
The name must correspond to the one-letter-code of the modified amino acid
and the numeric value must represent the mass that should be added to the
original amino accid mass, default: Carbamidomethyl modifications =
c(C = 57.02146). Use Nterm or Cterm as names for modifications that should
be added to the amino respectively carboxyl-terminus.}

\item{variable_modifications}{A named numeric vector of variable modifications.
Depending on the maximum number of modifications (\code{max_mods}), all possible
combinations are returned.}

\item{max_mods}{A numeric indicating the maximum number of variable modifications
allowed on the sequence at once. Does not include fixed modifications.
Default value is positive infinity.}

\item{neutralLoss}{list, it has to have two named elments, namely water and
ammonia that contain a character vector which type of neutral loss should be
calculated. Currently neutral loss on the C terminal "Cterm", at the amino
acids c("D", "E", "S", "T") for "water" (shown with an ⁠_⁠) an
d c("K", "N", "Q", "R") for "ammonia" (shown with an *) are supported.}

\item{verbose}{logical(1). If TRUE (default) the used modifications are printed.}
}
\value{
A named list containing dataframes. The dataframes show all the
ions produced by fragmentation with all possible combinations of modifications.
The used modifications are displayed in the element names of the list. Amino
acids within brackets thus include a variable modification. Fixed modifications
are not displayed in the names.
}
\description{
Calculate ions produced by fragmentation, including possible modifications
}
\examples{
calculateFragments2(sequence = "ARGSHKATC", type = c("b", "y"), z = 1, 
fixed_modifications = c(C = 57), variable_modifications = c(S = 79, Y = 79, T = 79),
max_mods = 2)
}
