% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterPSMs.R
\name{filterPSMs}
\alias{filterPSMs}
\title{Filter out unreliable PSMs.}
\usage{
filterPSMs(
  x,
  decoy = "isDecoy",
  rank = "rank",
  accession = "DatabaseAccess",
  spectrumID = "spectrumID",
  verbose = TRUE
)
}
\arguments{
\item{x}{A \code{DataFrame} containing PSMs.}

\item{decoy}{The column name defining whether entries match the
decoy database. Default is \code{"isDecoy"}. The column should be a
\code{logical} and only PSMs holding a \code{FALSE} are
retained. Ignored is set to \code{NULL}.}

\item{rank}{The column name holding the rank of the PSM. Default
is \code{"rank"}. This column should be a \code{numeric} and only PSMs
having rank equal to 1 are retained. Ignored is set to \code{NULL}.}

\item{accession}{The column name holding the protein (groups)
accession. Default is \code{"DatabaseAccess"}. Ignored is set to
\code{NULL}.}

\item{spectrumID}{The name of the spectrum identifier
column. Default is \code{spectrumID}.}

\item{verbose}{\code{logical(1)} setting the verbosity flag.}
}
\value{
A new \code{DataFrame} with filtered out peptides and with the
same columns as the input \code{x}.
}
\description{
A function to filter out PSMs matching to the decoy database, of
rank greater than one and matching non-proteotypic peptides.
}
\details{
The PSMs should be stored in a \code{DataFrame} (or \code{data.frame} or
\code{tibble}) such as those produced by \code{\link[=readPSMs]{readPSMs()}}. Note that
this function should be used before reducing the PSM table.
}
\examples{
f <- msdata::ident(full.names = TRUE, pattern = "TMT")
basename(f)
id <- readPSMs(f)
filterPSMs(id)
}
\author{
Laurent Gatto
}
