% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments-utils.R
\name{getAminoAcids}
\alias{getAminoAcids}
\title{Amino acids}
\usage{
getAminoAcids()
}
\value{
\code{data.frame}
}
\description{
Returns a \code{data.frame} of amino acid properties: \code{AA},
\code{ResidueMass}, \code{Abbrev3}, \code{ImmoniumIonMass}, \code{Name},
\code{Hydrophobicity}, \code{Hydrophilicity}, \code{SideChainMass}, \code{pK1}, \code{pK2}
and \code{pI}.
}
\examples{

getAminoAcids()
}
\author{
Laurent Gatto
}
