% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments-utils.R
\name{getAtomicMass}
\alias{getAtomicMass}
\title{Atomic mass.}
\usage{
getAtomicMass()
}
\value{
A named \code{double}.
}
\description{
Returns a \code{double} of used atomic mass.
}
\examples{
getAtomicMass()
}
\author{
Sebastian Gibb
}
