% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mzR.R
\name{as}
\alias{as}
\alias{as.data.frame.mzRident}
\title{Coerce PSM data to a \code{data.frame}}
\arguments{
\item{from}{An object of class \code{mzRident} defined in the \code{mzR}
package.}
}
\value{
A \code{data.frame}
}
\description{
A function to convert the identification data contained in an
\code{mzRident} object (as defined in the \code{mzR} package) to a
\code{data.frame}. Each row represents a scan, which can however be
repeated several times if the PSM matches multiple proteins and/or
contains two or more modifications.
}
\examples{
## find path to an mzIdentML file
f <- msdata::ident(full.names = TRUE, pattern = "TMT")
basename(f)

library("mzR")
x <- openIDfile(f)
x
as(x, "data.frame")
}
\author{
Laurent Gatto
}
