% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpectraWithFragments.R
\name{plotSpectraWithFragments}
\alias{plotSpectraWithFragments}
\title{Plots an MS2 Spectrum with Fragments}
\usage{
plotSpectraWithFragments(
  x,
  y,
  xlab = "m/z",
  ylab = "intensity",
  type = "h",
  xlim = numeric(),
  ylim = numeric(),
  main = y,
  col = "#00000080",
  labelCex = 1,
  labelSrt = 0,
  labelAdj = NULL,
  labelPos = NULL,
  labelOffset = 0.5,
  axes = TRUE,
  frame.plot = axes,
  ppm = 20,
  tolerance = 0,
  matchCol = "#80B1D3",
  matchLwd = 1,
  matchLty = 1,
  matchPch = 16,
  ...
)
}
\arguments{
\item{x}{An instance of class \code{Spectra}.}

\item{y}{\code{character(1)} with a peptides sequence.}

\item{xlab}{character(1)\verb{with the label for the x-axis (by default}xlab = "m/z"`).}

\item{ylab}{\code{character(1)} with the label for the y-axis (by
default \code{ylab = "intensity"}).}

\item{type}{\code{character(1)} specifying the type of plot. See
\code{\link[=plot.default]{plot.default()}} for details. Defaults to \code{type = "h"} which
draws each peak as a line.}

\item{xlim}{\code{numeric(2)} defining the x-axis limits. The range of
m/z values are used by default.}

\item{ylim}{\code{numeric(2)} defining the y-axis limits. The range of
intensity values are used by default.}

\item{main}{\code{character(1)} with the title for the plot. By default
the peptide sequence \code{y} is used.}

\item{col}{color to be used to draw the peaks. Should be either of
length 1, or equal to the number of spectra (to plot each
spectrum in a different color) or be a \code{list} with colors for
each individual peak in each spectrum.}

\item{labelCex}{\code{numeric(1)} giving the amount by which the text
should be magnified relative to the default. See parameter
\code{cex} in \code{\link[=par]{par()}}.}

\item{labelSrt}{\code{numeric(1)} defining the rotation of the
label. See parameter \code{srt} in \code{\link[=text]{text()}}.}

\item{labelAdj}{see parameter \code{adj} in \code{\link[=text]{text()}}.}

\item{labelPos}{see parameter \code{pos} in \code{\link[=text]{text()}}.}

\item{labelOffset}{see parameter \code{offset} in \code{\link[=text]{text()}}.}

\item{axes}{\code{logical(1)} whether (x and y) axes should be drawn.}

\item{frame.plot}{\code{logical(1)} whether a box should be drawn
around the plotting area.}

\item{ppm}{m/z relative acceptable difference (in ppm) for peaks
to be considered matching (see \code{\link[MsCoreUtils:matching]{MsCoreUtils::common()}} for more
details).}

\item{tolerance}{absolute acceptable difference of m/z values for
peaks to be considered matching (see \code{\link[MsCoreUtils:matching]{MsCoreUtils::common()}}
for more details).}

\item{matchCol}{color for matching peaks.}

\item{matchLwd}{line width (\code{lwd}) to draw matching peaks. See
\code{\link[=par]{par()}} for more details.}

\item{matchLty}{line type (\code{lty}) to draw matching peaks. See
\code{\link[=par]{par()}} for more details.}

\item{matchPch}{point character (\code{pch}) to label matching
peaks. Defaults to \code{matchPch = 16}, set to \code{matchPch = NA} to
disable. See \code{\link[=par]{par()}} for more details}

\item{...}{additional parameters to be passed to \code{\link[=plot.default]{plot.default()}}.}
}
\value{
Used to create a plot.
}
\description{
Ths
}
\examples{

sp <- DataFrame(msLevel = 2L, rtime = 2345, sequence = "SIGFEGDSIGR")
sp$mz <- list(c(100.048614501953, 110.069030761719, 112.085876464844,
                117.112571716309, 158.089569091797, 163.114898681641,
                175.117172241211, 177.098587036133, 214.127075195312,
                232.137542724609, 233.140335083008, 259.938415527344,
                260.084167480469, 277.111572265625, 282.680786132812,
                284.079437255859, 291.208282470703, 315.422576904297,
                317.22509765625, 327.2060546875, 362.211944580078,
                402.235290527344, 433.255004882812, 529.265991210938,
                549.305236816406, 593.217041015625, 594.595092773438,
                609.848327636719, 631.819702148438, 632.324035644531,
                632.804931640625, 640.8193359375, 641.309936523438,
                641.82568359375, 678.357238769531, 679.346252441406,
                688.291259765625, 735.358947753906, 851.384033203125,
                880.414001464844, 881.40185546875, 919.406433105469,
                938.445861816406, 1022.56658935547, 1050.50415039062,
                1059.82800292969, 1107.52734375, 1138.521484375,
                1147.51538085938, 1226.056640625))
sp$intensity <- list(c(83143.03, 65473.8, 192735.53, 3649178.5,
                       379537.81, 89117.58, 922802.69, 61190.44,
                       281353.22, 2984798.75, 111935.03, 42512.57,
                       117443.59, 60773.67, 39108.15, 55350.43,
                       209952.97, 37001.18, 439515.53, 139584.47,
                       46842.71, 1015457.44, 419382.31, 63378.77,
                       444406.66, 58426.91, 46007.71, 58711.72,
                       80675.59, 312799.97, 134451.72, 151969.72,
                       3215457.75, 1961975, 395735.62, 71002.98,
                       69405.73, 136619.47, 166158.69, 682329.75,
                       239964.69, 242025.44, 1338597.62, 50118.02,
                       1708093.12, 43119.03, 97048.02, 2668231.75,
                       83310.2, 40705.72))
sp <- Spectra::Spectra(sp)
plotSpectraWithFragments(sp, sp$sequence, labelPos = 3)
}
\author{
Johannes Rainer, Sebastian Gibb, Laurent Gatto
}
