% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPSMs.R
\name{readPSMs}
\alias{readPSMs}
\title{Import peptide-spectrum matches}
\usage{
readPSMs(files, parser = c("mzR", "mzID"), BPPARAM = SerialParam())
}
\arguments{
\item{files}{A \code{character} of \code{mzid} files.}

\item{parser}{\code{character(1)} defining the parser to be used to
read the \code{mzIdentML} files. One of \code{"mzR"} (default) or
\code{"mzID"}.}

\item{BPPARAM}{an object inheriting from \link{BiocParallelParam} to
control parallel processing. The default value is
\code{SerialParam()} to read files in series.}
}
\value{
A \code{DataFrame} containing the PSMs stored in the \code{mzId}
files.
}
\description{
Reads a set of \code{mzId} files containing PSMs and returns the PSMs
as a \code{DataFrame}.
}
\details{
This function uses parsers provided by the \code{mzR} or \code{mzID}
packages to read the \code{mzIdentML} data. See the vignette for some
apparent differences in their outputs.
}
\examples{
f <- msdata::ident(full.names = TRUE, pattern = "TMT")
basename(f)

## mzR parser (default)
readPSMs(f)

## mzID parser
readPSMs(f, parser = "mzID")
}
\author{
Laurent Gatto
}
