% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spBoekweg}
\alias{spBoekweg}
\title{Example mass spectrometry data}
\format{
A \code{Spectra} object with 4,766 spectra stored in memory using
\code{MsBackendMemory()}. The spectra contain standard metadata
variables including:

\describe{
\item{\code{msLevel}}{MS level (1 or 2)}
\item{\code{rtime}}{Retention time (seconds)}
\item{\code{precursorMz}}{Precursor m/z (MS2 only)}
\item{\code{precursorCharge}}{Precursor charge (MS2 only)}
\item{\code{dataOrigin}}{Source file name}
\item{\code{spectrumId}}{Spectrum identifier}
}
}
\usage{
data(spBoekweg)
}
\description{
\code{spBoekweg} is a \code{Spectra} object containing 4,766 mass spectra
(MS1 and MS2) that are linked to peptide identifications in the
\code{psmBoekweg} dataset.
}
\details{
This is an example subset of the data available through
\code{?MsDataHub::Boekweg2022()}.

The spectra can be linked to the \code{psmBoekweg} dataset using a
common key based on the file name and scan number. See the vignette
\code{validatePSM} for an example.
}
\examples{
data(spBoekweg)
spBoekweg

## Number of spectra
length(spBoekweg)

## MS levels
table(spBoekweg$msLevel)

## Retention time range (in seconds)
range(spBoekweg$rtime)

## Plot base peak chromatogram
plot(spBoekweg$rtime, spBoekweg$basePeakIntensity,
     xlab = "Retention time (s)", ylab = "Base peak intensity",
     type = "l")

## Link to PSM data using a common key
data(psmBoekweg)
psmBoekweg$pkey <- paste0(
    basename(psmBoekweg$filename),
    sub("^.+scan=", "::", psmBoekweg$scannr)
)
spBoekweg$pkey <- paste0(
    basename(spBoekweg$dataOrigin),
    sub("^.+scan=", "::", spBoekweg$spectrumId)
)

## Find spectra with identifications
sum(spBoekweg$pkey \%in\% psmBoekweg$pkey)
}
\seealso{
\itemize{
\item \link{psmBoekweg} for the corresponding \code{PSM} object with peptide
identifications.
\item \code{\link[Spectra:Spectra]{Spectra::Spectra()}} for the Spectra class from the Spectra
package.
\item The \code{inst/scripts/make_BoekwegData.R} script documenting data
creation.
}
}
\keyword{datasets}
