% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectedComponent-class.R
\name{ConnectedComponents}
\alias{ConnectedComponents}
\alias{ConnectedComponents-class}
\title{Connected components}
\usage{
ConnectedComponents(object)
}
\arguments{
\item{object}{An adjacency matrix class \code{Matrix} or an instance of
class \code{PSM}.}
}
\description{
Connected components are a useful representation when exploring
identification data. They represent the relation between proteins
(the connected components) and how they form groups of proteins
defined by shared peptides.

Connected components are stored as \code{Connectedcomponents} that can
be generated using the \code{ConnectedComponents()} function.
}
\examples{

## --------------------------------
## From an adjacency matrix
## --------------------------------
library(Matrix)
adj <- sparseMatrix(i = c(1, 2, 3, 3), j = c(1, 2, 3, 4), x = 1,
       dimnames = list(paste0("Pep", 1:3),
                       paste0("Prot", 1:4)))
adj
cc <- ConnectedComponents(adj)
cc

## --------------------------------
## From an PSM object
## --------------------------------
f <- msdata::ident(full.names = TRUE, pattern = "TMT")
f

psm <- PSM(f) |>
       filterPsmDecoy() |>
       filterPsmRank()

cc <- ConnectedComponents(psm)
cc
}
