% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterPSMs.R
\name{filterPSMs}
\alias{filterPSMs}
\alias{filterPsmDecoy}
\alias{filterPsmRank}
\alias{filterPsmNonProteotypic}
\alias{filterPsmBestScore}
\alias{filterPsmMods}
\title{Filter out unreliable PSMs.}
\usage{
filterPSMs(
  x,
  decoy = "isDecoy",
  rank = "rank",
  accession = "DatabaseAccess",
  spectrumID = "spectrumID",
  verbose = TRUE
)

filterPsmDecoy(x, decoy = "isDecoy", verbose = TRUE)

filterPsmRank(x, rank = "rank", verbose = TRUE)

filterPsmNonProteotypic(
  x,
  accession = "DatabaseAccess",
  spectrumID = "spectrumID",
  verbose = TRUE
)

filterPsmBestScore(x, score, spectrumID = "spectrumID", verbose = TRUE)

filterPsmMods(x, mod = "modName", verbose = TRUE)
}
\arguments{
\item{x}{An instance of class \code{PSM}.}

\item{decoy}{\code{character(1)} with the column name specifying
whether entries match the decoy database or not. Default is
\code{"isDecoy"}. The column should be a \code{logical} and only PSMs
holding a \code{FALSE} are retained. Filtering is ignored if set to
\code{NULL}.}

\item{rank}{\code{character(1)} with the column name holding the rank
of the PSM. Default is \code{"rank"}. This column should be a
\code{numeric} and only PSMs having rank equal to 1 are
retained. Filtering is ignored if set to \code{NULL}.}

\item{accession}{\code{character(1)} with the column name holding the
protein (groups) accession. Default is
\code{"DatabaseAccess"}. Filtering is ignored if set to \code{NULL}.}

\item{spectrumID}{\code{character(1)} with the name of the spectrum
identifier column. Default is \code{spectrumID}. Filtering is
ignored if set to \code{NULL}.}

\item{verbose}{\code{logical(1)} setting the verbosity flag.}

\item{score}{\code{character(1)} with the name of a score column to be
used. Filtering is ignored if set to \code{NULL}.}

\item{mod}{\code{character(1)} with the column name holding the
modification name. Default is \code{"modName"}. Filtering is
ignored if set to \code{NULL}.}
}
\value{
A new filtered \code{PSM} object with the same columns as the
input \code{x}.
}
\description{
Functions to filter out PSMs matching. The PSMs should be stored
in a \code{PSM} such as those produced by \code{\link[=PSM]{PSM()}}.

\itemize{
\item \code{filterPsmDecoy()} filters out decoy PSMs, i.e. those annotated
as \code{isDecoy}.
}

\itemize{
\item \code{filterPsmRank()} filters out PSMs of rank > 1.
}

\itemize{
\item \code{filterPsmNonProteotypic()} filters out non-proteotypic PSMs,
i.e. those that match multiple proteins.
}

\itemize{
\item \code{filterPsmBestScore()} filters out PSMs but those with he
highest score when matching the same spectrum identifier.
}

\itemize{
\item \code{filterPsmMods()} filters out PSMs that contain any PTM, as
defined by non-missing (i.e. non-\code{NA}) modification name.
}
}
\examples{
f <- msdata::ident(full.names = TRUE, pattern = "TMT")
basename(f)
id <- PSM(f)
filterPSMs(id)
}
\author{
Laurent Gatto
}
