% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fragments-label.R
\name{labelFragments}
\alias{labelFragments}
\title{labels MS2 Fragments}
\usage{
labelFragments(x, tolerance = 0, ppm = 20, what = c("ion", "mz"), ...)
}
\arguments{
\item{x}{An instance of class \code{Spectra} of length 1, containing a
spectra variable \code{"sequence"} with a \code{character(1)}
representing a valid peptide sequence.}

\item{tolerance}{absolute acceptable difference of m/z values for peaks to
be considered matching (see \code{\link[MsCoreUtils:matching]{MsCoreUtils::closest()}} for more details).}

\item{ppm}{m/z relative acceptable difference (in ppm) for peaks to be
considered matching (see \code{\link[MsCoreUtils:matching]{MsCoreUtils::closest()}} for more details).}

\item{what}{\code{character(1)}, one of \code{"ion"} (default) or \code{"mz"}, defining
whether labels should be fragment ions, , or their m/z values. If the
latter, then the m/z values are named with the ion labels.}

\item{...}{additional parameters (except \code{verbose}) passed to
\code{\link[=calculateFragments]{calculateFragments()}} to calculate fragment m/z values to be
added to the spectra in \code{x}.}
}
\value{
Return a \code{list()} of \code{character()} with fragment ion labels. The
elements are named after the peptide they belong to (variable
modifications included).
}
\description{
Creates a list of annotations based on \code{calculateFragments} results.
}
\examples{
library("Spectra")

## Load the PSM and spectra data
data("psmBoekweg")
data("spBoekweg")

## The spectra need a 'sequence' variable for `calculateFragments()`
## Make sure you can join both psms and spectra using `joinSpectraData`:
head(psmBoekweg$pkey <- paste0(basename(psmBoekweg$filename),
                               sub("^.+scan=", "::", psmBoekweg$scannr)))
head(spBoekweg$pkey <- paste0(basename(spBoekweg$dataOrigin),
                              sub("^.+scan=", "::", spBoekweg$spectrumId)))

sp <- Spectra::joinSpectraData(spBoekweg, psmBoekweg, by.x = "pkey")
sp <- sp[sp$msLevel == 2]

## Add carbamidomethylation or other modifications if need be
## See ?PTMods::addFixedModifications
(seq <- psmVariables(psmBoekweg)[["peptide"]])
(fdr <- psmVariables(psmBoekweg)[["fdr"]])

head(sp[["sequence"]] <- addFixedModifications(sp[[seq]]))

## The fragment ion labels
labelFragments(sp[1])

## The fragment mz labels
labelFragments(sp[1], what = "mz")

## Pass additional parameters to calculateFragments using a PTMods modified sequence
sp_mod <- sp[1]
sp_mod$sequence <- PTMods::addFixedModifications(sp_mod$sequence,
                                                  fixedModifications = c(Nterm = 5))
labelFragments(sp_mod, type = c("a", "b", "x", "y"))

## Annotate the spectum with the fragment labels
plotSpectra(sp[1], labels = labelFragments, labelPos = 3)

## By default used in `plotSpectraPTM()`.
plotSpectraPTM(sp[1])
}
\author{
Johannes Rainer, Guillaume Deflandre, Sebastian Gibb, Laurent Gatto
}
