% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacencyMatrix.R
\name{makeAdjacencyMatrix}
\alias{makeAdjacencyMatrix}
\title{Create an adjacency matrix.}
\usage{
makeAdjacencyMatrix(
  x,
  split = ";",
  peptides = "sequence",
  proteins = "DatabaseAccess",
  sparse = FALSE
)
}
\arguments{
\item{x}{Either an instance of class \code{PSM} or a \code{character}. See
example below for details.}

\item{split}{\code{character(1)} defining how to split the string of
protein identifiers.}

\item{peptides}{\code{character(1)} indicating the name of the variable
that defines peptides in the \code{PSM} object. Default is
\code{sequence}.}

\item{proteins}{\code{character(1)} indicating the name of the variable
that defines proteins in the \code{PSM} object. Default is
\code{DatanbaseAccess}.}

\item{sparse}{\code{logical(1)} defining is a sparse adjacency matrix
should be generated. Default is \code{FALSE}.}
}
\value{
A peptide-by-protein adjacency \code{matrix}.
}
\description{
This function created a peptide-by-protein adjacency matrix from a
\code{character} or an instance of class \code{PSM}.

The character is formatted as \code{x <- c("ProtA", "ProtB", "ProtA;ProtB", ...)}, as commonly encoutered in proteomics data
spreadsheets. It defines that the first peptide is mapped to
protein "ProtA", the second one to protein "ProtB", the third one
to "ProtA" and "ProtB", and so on. The resulting matrix contain
\code{length(x)} rows an as many columns as there are unique protein
idenifiers in \code{x}. The column are always named after the protein
idenifiers. If the protein identifier vector is named and the
names are unique, these are then used to name to matrix rows.
}
\examples{
## Protein vector without names
prots <- c("ProtA", "ProtB", "ProtA;ProtB")
makeAdjacencyMatrix(prots)

## Named protein vector
names(prots) <- c("pep1", "pep2", "pep3")
makeAdjacencyMatrix(prots)
}
\author{
Laurent Gatto
}
