% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacencyMatrix.R
\name{makeAdjacencyMatrix}
\alias{makeAdjacencyMatrix}
\title{Create an adjacency matrix.}
\usage{
makeAdjacencyMatrix(
  x,
  split = ";",
  peptides = "sequence",
  proteins = "DatabaseAccess",
  unique = FALSE
)
}
\arguments{
\item{x}{Either an instance of class \code{PSM} or a \code{character}. See
example below for details.}

\item{split}{\code{character(1)} defining how to split the string of
protein identifiers (using \code{\link[=strsplit]{strsplit()}}). Default is ";". If
\code{NULL}, splitting is ignored.}

\item{peptides}{\code{character(1)} indicating the name of the variable
that defines peptides in the \code{PSM} object. Default is
\code{sequence}.}

\item{proteins}{\code{character(1)} indicating the name of the variable
that defines proteins in the \code{PSM} object. Default is
\code{DatanbaseAccess}.}

\item{unique}{\code{logical(1)} defining if all peptides should should
be considered in the contingency matrix, or should duplicates
be ingored (and only first occurences taken into account). The
default is \code{FALSE} and all peptides names are made unique by
appending sequence numbers to duplicates (see
\code{\link[=make.unique]{make.unique()}}).}
}
\value{
A peptide-by-protein adjacency matrix.
}
\description{
This function created a peptide-by-protein adjacency matrix from a
\code{character} or an instance of class \code{PSM}.
}
\details{
The character is formatted as \code{x <- c("ProtA", "ProtB", "ProtA;ProtB", ...)}, as commonly encoutered in proteomics data
spreadsheets. It defines that the first peptide is mapped to
protein "ProtA", the second one to protein "ProtB", the third one
to "ProtA" and "ProtB", and so on. The resulting matrix contain
\code{length(x)} rows an as many columns as there are unique protein
idenifiers in \code{x}. The column are always named after the protein
idenifiers. If the protein identifier vector is named and the
names are unique, these are then used to name to matrix rows.

The adjacency matrix produced from a \code{PSM} object doesn't
represent a peptide-by-protein adjacency matrix, given that the
observations of a \code{PSM} object represent peptide-spectrum matches,
rather than peptides. It is possible to set the \code{unique} parameter
to return unique PSM occurences. Note however that this implicitly
assumes that the same peptides where matched to the same
protein(s), which is not explicitly verified.
}
\examples{
## Protein vector without names
prots <- c("ProtA", "ProtB", "ProtA;ProtB")
makeAdjacencyMatrix(prots)

## Named protein vector
names(prots) <- c("pep1", "pep2", "pep3")
makeAdjacencyMatrix(prots)

## From a PSM object
f <- msdata::ident(full.names = TRUE, pattern = "TMT")
psm <- filterPSMs(PSM(f))
psm
adj <- makeAdjacencyMatrix(psm)
dim(adj)
adj[1:10, 1:4]

## Drop duplicated peptides
adj <- makeAdjacencyMatrix(psm, unique = TRUE)
dim(adj)
adj[1:10, 1:4]
}
\author{
Laurent Gatto
}
