% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacencyMatrix.R
\name{makeAdjacencyMatrix}
\alias{makeAdjacencyMatrix}
\title{Create an adjacency matrix.}
\usage{
makeAdjacencyMatrix(
  x,
  split = ";",
  peptide = psmVariables(x)["peptide"],
  protein = psmVariables(x)["protein"]
)
}
\arguments{
\item{x}{Either an instance of class \code{PSM} or a \code{character}. See
example below for details.}

\item{split}{\code{character(1)} defining how to split the string of
protein identifiers (using \code{\link[=strsplit]{strsplit()}}). Default is ";". If
\code{NULL}, splitting is ignored.}

\item{peptide}{\code{character(1)} indicating the name of the variable
that defines peptides in the \code{PSM} object. Default is the
\code{peptide} PSM variable as defined in \code{\link[=psmVariables]{psmVariables()}}.}

\item{protein}{\code{character(1)} indicating the name of the variable
that defines proteins in the \code{PSM} object. Default is the
\code{peptide} PSM variable as defined in \code{\link[=psmVariables]{psmVariables()}}.Default
is \code{DatanbaseAccess}.}
}
\value{
A peptide-by-protein adjacency matrix.
}
\description{
This function created a peptide-by-protein adjacency matrix from a
\code{character} or an instance of class \code{PSM}.
}
\details{
The character is formatted as \code{x <- c("ProtA", "ProtB", "ProtA;ProtB", ...)}, as commonly encoutered in proteomics data
spreadsheets. It defines that the first peptide is mapped to
protein "ProtA", the second one to protein "ProtB", the third one
to "ProtA" and "ProtB", and so on. The resulting matrix contain
\code{length(x)} rows an as many columns as there are unique protein
idenifiers in \code{x}. The column are always named after the protein
idenifiers. If the protein identifier vector is named and the
names are unique, these are then used to name to matrix rows.
}
\examples{
## Protein vector without names
prots <- c("ProtA", "ProtB", "ProtA;ProtB")
makeAdjacencyMatrix(prots)

## Named protein vector
names(prots) <- c("pep1", "pep2", "pep3")
makeAdjacencyMatrix(prots)

## From a PSM object
f <- msdata::ident(full.names = TRUE, pattern = "TMT")
library(magrittr)
psm <- PSM(f) \%>\%
       filterPsmDecoy() \%>\%
       filterPsmRank()
psm
adj <- makeAdjacencyMatrix(psm)
dim(adj)
adj[1:10, 1:4]
## Peptides with rowSums > 1 match multiple proteins.
## Use filterPsmNonProteotypic() to filter these out.
table(rowSums(adj))
}
\author{
Laurent Gatto
}
