% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mzR.R
\name{as_data_frame}
\alias{as_data_frame}
\title{Coerce PSM data to a \code{data.frame}}
\usage{
as_data_frame(from)
}
\arguments{
\item{from}{An object of class \code{mzRident} as defined in the \code{mzR}
package.}
}
\value{
A \code{data.frame}
}
\description{
A function to convert the identification data contained in an
\code{mzRident} object (as defined in the \code{mzR} package) to a
\code{data.frame}. Each row represents a scan, which can however be
repeated several times if the PSM matches multiple proteins and/or
contains two or more modifications.
}
\details{
This function implements the same funtionality as \code{MSnbase::as(., "data.frame")}. It as been renamed and isn't been exported to
avoid conflicts and confusion. Ideally, the function in \code{MSnbase}
might be moved to \code{mzR} and this package would re-use that one.
}
\examples{
## find path to an mzIdentML file
f <- msdata::ident(full.names = TRUE, pattern = "TMT")
basename(f)

library("mzR")
x <- openIDfile(f)
x
PSMatch:::as_data_frame(x)
}
\author{
Laurent Gatto
}
