% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSM-class.R, R/reducePSMs.R
\name{PSM}
\alias{PSM}
\alias{PSM-class}
\alias{reducePSMs}
\alias{readPSMs}
\title{A class for peptide-spectrum matches results}
\usage{
PSM(files, parser = c("mzR", "mzID"), BPPARAM = SerialParam())

reducePSMs(x, k)
}
\arguments{
\item{files}{\code{character()} of \code{mzid} file names.}

\item{parser}{\code{character(1)} defining the parser to be used to
read the \code{mzIdentML} files. One of \code{"mzR"} (default) or
\code{"mzID"}.}

\item{BPPARAM}{an object inheriting from \link{BiocParallelParam} to
control parallel processing. The default value is
\code{SerialParam()} to read files in series.}

\item{x}{An instance of class \code{PSM}.}

\item{k}{A \code{vector} or \code{factor} of length equal to \code{nrow(x)} that
defines the primary key used to reduce \code{x}. This typically
corresponds to the spectrum identifier, often names
\code{spectrumID}.}
}
\value{
\code{PSM()} returns a \code{PSM} object containing the
peptide-spectrum matches stored in \code{files}.

\code{reducePSMs()} returns a reduced version of the \code{x} input.
}
\description{
The \code{PSM} class is a simple class to store and manipulate
peptide-spectrum matches. The class encapsulates PSM data as a
\code{DataFrame} (or more specifically a \code{DFrame}) with additional
lightweight metadata annotation.

There are two types of \code{PSM} objects:
\itemize{
\item Objects where each PSM is stored in a single data row. This also
holds for multiple matches to the same spectrum, be it different
peptide sequences or the same sequence with or without a
post-translational modification. In such cases, the spectrum
identifier column contains duplicated entries. Such objects are
typically created with the \code{PSM()} constructor starting from
\code{mzIdentML} files.
\item Reduced object where the spectrum identifiers are unique keys
within the PSM table. Matches to the same scan/spectrum are
merged into a single data row. Reduced \code{PSM} object are created
with the \code{reducePSMs()} function. See examples below.
}

The constructor uses parsers provided by the \code{mzR} or \code{mzID}
packages to read the \code{mzIdentML} data. See the vignette for some
apparent differences in their outputs.
}
\examples{
f <- msdata::ident(full.names = TRUE, pattern = "TMT")
basename(f)

## mzR parser (default)
psm <- PSM(f)
psm

## mzID parser
PSM(f, parser = "mzID")

i <- which(duplicated(psm$spectrumID))[1:2]
i <- which(psm$spectrumID \%in\% psm$spectrumID[i])
psm2 <- psm[i, ]

## Peptide sequence CIDRARHVEVQIFGDGKGRVVALGERDCSLQRR with
## Carbamidomethyl modifications at positions 1 and 28.
DataFrame(psm2[, c("sequence", "spectrumID", "modName", "modLocation")])

rpsm2 <- reducePSMs(psm2, psm2$spectrumID)
rpsm2
DataFrame(rpsm2[, c("sequence", "spectrumID", "modName", "modLocation")])
}
