% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSM-class.R, R/reducePSMs.R
\name{PSM}
\alias{PSM}
\alias{PSM-class}
\alias{reducePSMs}
\alias{readPSMs}
\alias{PSM,character}
\alias{PSM,data.frame}
\alias{PSM,PSM}
\alias{reduced}
\alias{reduced<-}
\alias{psmVariables}
\alias{psmVariables<-}
\title{A class for peptide-spectrum matches}
\usage{
PSM(
  x,
  spectrum = NA,
  peptide = NA,
  protein = NA,
  decoy = NA,
  rank = NA,
  parser = c("mzR", "mzID"),
  BPPARAM = SerialParam()
)

reduced(object, spectrum = psmVariables(object)["spectrum"])

reduced(object) <- value

psmVariables(object, which = "all")

psmVariables(object) <- value

reducePSMs(object, k = object[[psmVariables(object)["spectrum"]]])
}
\arguments{
\item{x}{\code{character()} of \code{mzid} file names or an instance of
class \code{PSM}.}

\item{spectrum}{\code{character(1)} that defines a spectrum in the PSM
data. Default are \code{"spectrumID"} (mzR parser) or
\code{"spectrumid"} (mzID parser). It is also used to calculate the
reduced state.}

\item{peptide}{\code{character(1)} that defines a peptide in the PSM
data. Detaults are \code{"spequence"} (mzR parser) or \code{"pepSeq"}
(mzID parser).}

\item{protein}{\code{character(1)} that defines a protein in the PSM
data. Detaults are \code{"DatabaseAccess"} (mzR parser) or
\code{"accession"} (mzID parser).}

\item{decoy}{\code{character(1)} that defines a decoy hit in the PSM
data. Detaults are \code{"isDecoy"} (mzR parser) or \code{"isdecoy"}
(mzID parser).}

\item{rank}{\code{character(1)} that defines the rank of the petide
spetrum match in the PSM data. Default is \code{"rank"}.}

\item{parser}{\code{character(1)} defining the parser to be used to
read the \code{mzIdentML} files. One of \code{"mzR"} (default) or
\code{"mzID"}.}

\item{BPPARAM}{an object inheriting from \link{BiocParallelParam} to
control parallel processing. The default value is
\code{SerialParam()} to read files in series.}

\item{object}{An instance of class \code{PSM}.}

\item{value}{new value to be passed to setter.}

\item{which}{\code{character()} with the PSM variable name to
retrieve. If \code{"all"} (default), all named variables are
returned. See \code{\link[=PSM]{PSM()}} for valid PSM variables.}

\item{k}{A \code{vector} or \code{factor} of length equal to \code{nrow(x)} that
defines the primary key used to reduce \code{x}. This typically
corresponds to the spectrum identifier. The defauls is to use
the spectrum PSM variable.}
}
\value{
\code{PSM()} returns a \code{PSM} object containing the
peptide-spectrum matches stored in \code{files}.

\code{reducePSMs()} returns a reduced version of the \code{x} input.
}
\description{
The \code{PSM} class is a simple class to store and manipulate
peptide-spectrum matches. The class encapsulates PSM data as a
\code{DataFrame} (or more specifically a \code{DFrame}) with additional
lightweight metadata annotation.

There are two types of \code{PSM} objects:
\itemize{
\item Objects with duplicated spectrum identifiers. This holds for
multiple matches to the same spectrum, be it different peptide
sequences or the same sequence with or without a
post-translational modification. Such objects are typically
created with the \code{PSM()} constructor starting from \code{mzIdentML}
files.
\item Reduced object where the spectrum identifiers (or any equivalent
column - see example below) are unique keys within the PSM
table. Matches to the same scan/spectrum in each raw data file
are merged into a single data row. Reduced \code{PSM} object are
created with the \code{reducePSMs()} function. See examples below.
}

Objects can be checked for their reduced state with the
\code{reduced()} function which returns \code{TRUE} for reduced instances,
\code{FALSE} when the spectrum identifiers are duplicated, or NA when
unknown. The flag can also be set explicitly with the
\verb{reduced()<-} setter.
\itemize{
\item The constructor can also be initialise some variables needed for
downstream processing, notably filtering (See
\code{\link[=filterPSMs]{filterPSMs()}}). These variables can be extracted with the
\code{\link[=psmVariables]{psmVariables()}} function. They represent the columns in the PSM
table that identify spectra, peptides, proteins, decoy peptides
and hit ranks.
}

The constructor uses parsers provided by the \code{mzR} or \code{mzID}
packages to read the \code{mzIdentML} data. See the vignette for some
apparent differences in their outputs.
}
\examples{

## ---------------------------------
## Example with a single mzid file
## ---------------------------------

f <- msdata::ident(full.names = TRUE, pattern = "TMT")
basename(f)

## mzR parser (default)
psm <- PSM(f)
psm

## PSM variables
psmVariables(psm)

## mzID parser
psm_mzid <- PSM(f, parser = "mzID")
psm_mzid

## different PSM variables
psmVariables(psm_mzid)

## Reducing the PSM data
(i <- which(duplicated(psm$spectrumID))[1:2])
(i <- which(psm$spectrumID \%in\% psm$spectrumID[i]))
psm2 <- psm[i, ]
reduced(psm2)

## Peptide sequence CIDRARHVEVQIFGDGKGRVVALGERDCSLQRR with
## Carbamidomethyl modifications at positions 1 and 28.
DataFrame(psm2[, c("sequence", "spectrumID", "modName", "modLocation")])
reduced(psm2) <- FALSE
reduced(psm2)

## uses by default the spectrum PSM variable, as defined during
## the construction - see psmVariables()
rpsm2 <- reducePSMs(psm2)
rpsm2
DataFrame(rpsm2[, c("sequence", "spectrumID", "modName", "modLocation")])
reduced(rpsm2)

## ---------------------------------
## Multiple mzid files
## ---------------------------------

library(rpx)
PXD022816 <- PXDataset("PXD022816")
PXD022816

(mzids <- pxget(PXD022816, grep("mzID", pxfiles(PXD022816))[1:2]))
psm <- PSM(mzids)
psm
psmVariables(psm)

## Here, spectrum identifiers are repeated accross files
psm[grep("scan=20000", psm$spectrumID), "spectrumFile"]

## Let's create a new primary identifier composed of the scan
## number and the file name
psm$pkey <- paste(sub("^.+Task\\\\\\\\", "", psm$spectrumFile),
                  sub("^.+scan=", "", psm$spectrumID),
                  sep = "::")
head(psm$pkey)

## the PSM is not reduced
reduced(psm, "pkey")
DataFrame(psm[6:7, ])

## same sequence, same spectrumID, same file
psm$sequence[6:7]
psm$pkey[6:7]

## different modification locations
psm$modLocation[6:7]

## here, we need to *explicitly* set pkey to reduce
rpsm <- reducePSMs(psm, psm$pkey)
rpsm
reduced(rpsm, "pkey")

## the two rows are now merged into a single one; the distinct
## modification locations are preserved.
(i <- which(rpsm$pkey == "QEP2LC6_HeLa_50ng_251120_01-calib.mzML::12894"))
DataFrame(rpsm[i, c("sequence", "pkey", "modName", "modLocation")])
}
