% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMatch-package.R
\docType{package}
\name{PSMatch}
\alias{PSMatch}
\title{PSMatch: Handling and Managing Peptide Spectrum Matches}
\description{
The PSMatch package offers functionality to load, manage and
analyse Peptide Spectrum Matches as generated in mass
spectrometry-based proteomics. The three main objects and concepts
that are proposed in this package are described below, and are
aimed to proteomics practitioners to explore and understand their
identification data better.
}
\section{PSM objects}{


As mentioned in the \code{\link[=PSM]{PSM()}} manual page, The \code{PSM} class is a
simple class to store and manipulate peptide-spectrum matches. The
class encapsulates PSM data as a DataFrame (or more specifically a
\code{DFrame}) with additional lightweight metadata annotation. PSM
objects are typically creatd from XML-based mzID files or
\code{data.frames} imported from spreadsheets. It is them possible to
apply widely used filters (such as removal of decoy hits, PSMs of
rank > 1, ...) as described in \code{\link[=filterPSMs]{filterPSMs()}}.
}

\section{Adjacency matrices}{


PSM data, as produced by all proteomics search engines, is exported
as a table like structure where PSM are documented along the rows
by variables such as identification scores, peptides sequences,
modifications and the protein which that peptide comes from. There
is always a level of ambiguity in such data, as peptides can be
mapped to mutliple proteins; they are then called shared peptides,
as opposed to unique peptides.

One convenient way to store the relation between peptides and
proteins is as a peptide-by-protein adjacency matrix. Such matrices
can be generated from PSM object or vectors using the
\code{\link[=makeAdjacencyMatrix]{makeAdjacencyMatrix()}}.

The \code{\link[=describePeptides]{describePeptides()}} and \code{\link[=describeProteins]{describeProteins()}} functions are
also helpful is tally the number of unique and shared peptides and
the number of proteins composed of unique or shared peptides, or a
combination thereof.
}

\section{Connected Components}{


Once we model the peptide-to-protein relations explicitly, it
becomes possible to perform computations on the proteins that are
grouped by the peptides they share. These groups are mathematically
defined as connected components, which are implemented as
\code{\link[=ConnectedComponcents]{ConnectedComponcents()}} objects.
}

\section{Fragment ions}{


The package also provides functionality to calculate ions produced
by the fragmentation of a peptides (see \code{\link[=calculateFragments]{calculateFragments()}}) and
annotated MS2 \code{\link[Spectra:Spectra]{Spectra::Spectra()}} objects (see \code{\link[=addFragments]{addFragments()}}).
}

