% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIdentifications.R
\name{readPSMs}
\alias{readPSMs}
\title{Import peptide-spectrum matches}
\usage{
readPSMs(files, backend = c("mzR", "mzID"))
}
\arguments{
\item{files}{A \code{character} of \code{mzid} files.}

\item{backend}{\code{character(1)} defining the parser to be used to
read the \code{mzIdentML} files. One of \code{"mzR"} (default) or
\code{"mzID"}.}
}
\value{
A \code{DataFrame} containing the PSMs stored in the \code{mzId}
files.
}
\description{
Reads as set of \code{mzId} files containing PSMs and returns the PSMs
as a \code{DataFrame}.
}
\details{
This function uses parsers provided by the \code{mzR} or \code{mzID}
packages to read the \code{mzIdentML} data
}
\examples{
f <- msdata::ident(full.names = TRUE, pattern = "TMT")
basename(f)
readPSMs(f)
}
\author{
Laurent Gatto
}
